requirements.txt

streamlit
python-dotenv
smartapi-python
streamlit-autorefresh
pandas
